// src/nodecg/generated/playerdata.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Playerdata {
	playerData: {
		player1: Player;
		player2: Player;
		[k: string]: unknown;
	};
	[k: string]: unknown;
}
export interface Player {
	id: number;
	name: string;
	score: number;
	initialSeedNum: number;
	gamerTag: string;
	prefix?: string | null;
	externalUsername?: string | null;
	bio?: string | null;
	birthday?: string | null;
	city?: string | null;
	country?: string | null;
	winRate?: number;
	paginatedSetsDetails?: SetDetails[] | null;
	recentStandings?: Standing[] | null;
	rankings?: unknown[] | null;
	[k: string]: unknown;
}
export interface SetDetails {
	identifier: string;
	phasename: string;
	Round: string;
	player1name: string | null;
	player1score?: number | null;
	player2name?: string | null;
	player2score: number | null;
	[k: string]: unknown;
}
export interface Standing {
	tournamentName: string;
	standing: number;
	numEntrants: number;
	[k: string]: unknown;
}